refactor: Impl 제거, 패키지명 변경
